Drop commons-exec for forking processes
Motivation:

commons-exec is a legacy library and kind-off classpath dead weight (gatling-enterprise-plugin-commons is in the bundle’s runtime classpath, hence commons-exec is too). It's only there because the Fork class was forked from scala-maven-plugin to be compatible with pre-Java 7 world and it’s considered to be dropped there too.

Modification:

We can simply use ProcessBuilder instead.

Result:

* one less dependency, all the more a legacy one
* less dead weight in the bundle's classpath